improve py scripts...
