+++ e2e_test/s3/json_file.py
@@ -7,7 +7,6 @@ from minio import Minio
 import psycopg2
 import random
 
-
 def do_test(client, config, N,  prefix):
     conn = psycopg2.connect(
         host="localhost",
@@ -18,8 +17,6 @@ def do_test(client, config, N,  prefix):
 
     # Open a cursor to execute SQL statements
     cur = conn.cursor()
-
-    # Execute a SELECT statement
     cur.execute(f'''CREATE TABLE s3_test_jsonfile(
         id int,
         name TEXT,
